+++ spec/integration/resource/api600/synergy/logical_interconnect/create_spec.rb
@@ -1,6 +1,6 @@
 require 'spec_helper'
 
-klass = OneviewSDK::API500::Synergy::LogicalInterconnect
+klass = OneviewSDK::API600::Synergy::LogicalInterconnect
 RSpec.describe klass, integration: true, type: CREATE, sequence: seq(klass) do
   # Cannot create individually
 end